chore: Login 페이지 -> index 페이지로 로직 이동
